refactor: 폴더위치, 이미지 경로 next에 맞게 수정
