+++ src/pages/index.tsx
@@ -8,7 +8,6 @@ function MainPage() {
   return (
     <>
       <Header />
-      <Input />
     </>
   );
 }
