+++ src/pages/shared.tsx
@@ -1,7 +1,7 @@
-import Footer from "../../components/Footer/Footer";
-import Header from "../../components/Header/Header";
-import HeroContent from "../../components/HeroContent/HeroContent";
-import MainContent from "../../components/MainContent";
+import Footer from "../components/Footer/Footer";
+import Header from "../components/Header/Header";
+import HeroContent from "../components/HeroContent/HeroContent";
+import MainContent from "../components/MainContent";
 import styled from "styled-components";
 
 const MainContainer = styled.div`
